+++ src/reaction_network.jl
@@ -327,6 +327,22 @@ function unique_symbol_check(syms)
     nothing
 end
 
+# takes a ModelingToolkit declaration macro like @parameters and returns an expression
+# that calls the macro and then scalarizes all the symbols created into a vector of Nums
+function scalarize_macro(nonempty, ex, name)
+    namesym = gensym(name)
+    if nonempty
+        symvec = gensym()
+        ex = quote
+            $symvec = $ex
+            $namesym = reduce(vcat, Symbolics.scalarize($symvec))
+        end
+    else
+        ex = :($namesym = Num[])
+    end
+    ex, namesym
+end
+
 # Function for creating a ReactionSystem structure (used by the @reaction_network macro).
 function make_reaction_system(ex::Expr; name = :(gensym(:ReactionSystem)))
 
@@ -379,21 +395,24 @@ function make_reaction_system(ex::Expr; name = :(gensym(:ReactionSystem)))
     sexprs = get_sexpr(species_extracted, options; iv_symbols = ivs)
     vexprs = haskey(options, :variables) ? options[:variables] : :()
     pexprs = get_pexpr(parameters_extracted, options)
-    rxexprs = :($(make_ReactionSystem_internal)([], $tiv, Num[], Num[]; name = $(name),
-                                                spatial_ivs = $sivs))
-    foreach(speci -> push!(rxexprs.args[5].args, speci), species)
-    foreach(parameter -> push!(rxexprs.args[6].args, parameter), parameters)
+    ps, pssym = scalarize_macro(!isempty(parameters), pexprs, "ps")
+    vars, varssym = scalarize_macro(!isempty(variables), vexprs, "vars")
+    sps, spssym = scalarize_macro(!isempty(species), sexprs, "specs")
+    rxexprs = :(Catalyst.CatalystEqType[])
     for reaction in reactions
-        push!(rxexprs.args[3].args, get_rxexprs(reaction))
+        push!(rxexprs.args, get_rxexprs(reaction))
     end
 
     # Returns the rephrased expression.
     quote
-        $pexprs
+        $ps
         $ivexpr
-        $vexprs
-        $sexprs
-        $rxexprs
+        $vars
+        $sps
+
+        Catalyst.make_ReactionSystem_internal($rxexprs, $tiv, union($spssym, $varssym),
+                                              $pssym; name = $name,
+                                              spatial_ivs = $sivs)
     end
 end
 
@@ -615,7 +634,7 @@ end
 # stoichiometry. Recursion makes it able to handle weird cases like 2(X+Y+3(Z+XY)).
 function recursive_find_reactants!(ex::ExprValues, mult::ExprValues,
                                    reactants::Vector{ReactantStruct})
-    if typeof(ex) != Expr || (ex.head == :escape)
+    if typeof(ex) != Expr || (ex.head == :escape) || (ex.head == :ref)
         (ex == 0 || in(ex, empty_set)) && (return reactants)
         if any(ex == reactant.reactant for reactant in reactants)
             idx = findall(x -> x == ex, getfield.(reactants, :reactant))[1]